BASH SCRIPTING TO PARSE BIOLOGICAL DATA (SAMTOOLS)
EXERCISE

The aim of this exercise is simply to run a program across 3 sample data files to get the number of records they contain. Each line in the sample data files is a biological unit of data known as an alignment. 

QUESTION: How do we know how many records(alignments) are in each of the data files?

Note that, these files are not in a human readable format (BAM files). This means we need to use a program, like samtools, to process them to know the number of records(alignments) each file contains.
'samtools' - a suite of tools which can be used to process biological sequence data.

For this exercise, I used 'samtools view' to get a summary of the data in the sample files.
SAMPLE FILES;
- sample_10000_11000.bam
- sample_11000_12000.bam
- sample_12000_13000.bam

ADDITIONAL NOTES FOR BETTER UNDERSTANDING
> What does it mean to parse data?
Data parsing is the process of converting data from one format to another. It is used for data structuring, it is generally done to make the existing, often unstructured, unreadable data more comprehensible.

> Binary Alignment Map (BAM) file is the comprehensive raw data of genome sequencing; it consists of the lossless, compressed binary representation of the Sequence Alignment Map (SAM) file.
A SAM file (.sam) is a tab-delimited text file that contains sequence alignment data.
BAM files (.bam) contain the same information as SAM files, except they are in binary file format which is not readable by humans.

> The 'set' command is used to set or unset specific flags and settings (determines the behaviour of the script and helps in executing the tasks without any issues.) inside the shell environment. The 'set' command has the following command:
$ set [option] [argument]

> The '-s' option in Bash is used to check to see if a FILE exists and has a size greater than zero. It returns true and false values to indicate that the file is empty or has some data.